global actx init, clean up pyopencl usage (#1030)
