+++ scripts/quartz.sbatch.sh
@@ -24,4 +24,4 @@ export XDG_CACHE_HOME="${MIRGE_CACHE_ROOT}/xdg-cache"
 
 # Run application
 # -O: switch on optimizations
-srun -n $nproc python -O -m mpi4py ./vortex-mpi.py
+srun -n $nproc python -O -m mpi4py ./vortex.py
